requirements.txt
numpy
pandas
streamlit
openai
psycopg2-binary
sqlalchemy